+++ SofaKernel/framework/sofa/core/visual/DisplayFlags.cpp
@@ -134,13 +134,13 @@ std::istream& FlagTreeItem::read(std::istream &in)
 /*static*/
 void FlagTreeItem::create_parse_map(FlagTreeItem *root, std::map<std::string, bool, ci_comparison> &map)
 {
-    int sizeShow = root->m_showName.size();
-    int sizeHide = root->m_hideName.size();
-    for(int i=0; i<sizeShow; i++)
+    size_t sizeShow = root->m_showName.size();
+    size_t sizeHide = root->m_hideName.size();
+    for(size_t i=0; i<sizeShow; i++)
     {
         map[root->m_showName[i]] = false;
     }
-    for(int i=0; i<sizeHide; ++i)
+    for(size_t i=0; i<sizeHide; ++i)
     {
         map[root->m_hideName[i]] = false;
     }
@@ -159,12 +159,12 @@ void FlagTreeItem::read_recursive(FlagTreeItem *root, const std::map<std::string
     std::map<std::string,bool>::const_iterator iter_hide;
     for( iter = root->m_child.begin(); iter != root->m_child.end(); ++iter)
     {
-        int sizeShow = (*iter)->m_showName.size();
-        int sizeHide = (*iter)->m_hideName.size();
+        size_t sizeShow = (*iter)->m_showName.size();
+        size_t sizeHide = (*iter)->m_hideName.size();
 
         bool found = false;
 
-        for(int i=0; i<sizeHide; i++)
+        for(size_t i=0; i<sizeHide; i++)
         {
             iter_hide = parse_map.find((*iter)->m_hideName[i]);
             if( iter_hide != parse_map.end() )
@@ -184,7 +184,7 @@ void FlagTreeItem::read_recursive(FlagTreeItem *root, const std::map<std::string
                 }
             }
         }
-        for(int i=0; i<sizeShow; i++)
+        for(size_t i=0; i<sizeShow; i++)
         {
             iter_show = parse_map.find((*iter)->m_showName[i]);
             if( iter_show != parse_map.end() )
