+++ modules/SofaTopologyMapping/CenterPointTopologicalMapping.cpp
@@ -90,7 +90,7 @@ void CenterPointTopologicalMapping::updateTopologicalMappingTopDown()
             {
             case core::topology::HEXAHEDRAADDED:
             {
-                const unsigned int nbHexaAdded = ( static_cast< const HexahedraAdded *>( *changeIt ) )->getNbAddedHexahedra();
+                const size_t nbHexaAdded = ( static_cast< const HexahedraAdded *>( *changeIt ) )->getNbAddedHexahedra();
                 to_pstm->addPointsProcess(nbHexaAdded);
                 to_pstm->addPointsWarning(nbHexaAdded, true);
                 to_pstm->propagateTopologicalChanges();
